refactor(global): 테스트 디렉터리 구조 프로덕션 미러링 정리
